test error handling during reindex
